Move severityToStatus to Config and add severityEnum
