+++ love/src/components/Layout/CompactAlarm/CompactAlarm.jsx
@@ -1,14 +1,7 @@
 import React from 'react';
 import styles from './CompactAlarm.module.css';
 import StatusText from '../../GeneralPurpose/StatusText/StatusText';
-
-export const severityToStatus = {
-  0: 'unknown',
-  1: 'ok',
-  2: 'warning',
-  3: 'serious',
-  4: 'critical',
-};
+import { severityToStatus } from '../../../Config';
 
 export default function CompactAlarm({ name, severity, maxSeverity, acknowledged, muted, severityUpdateTimestamp }) {
   const severityStatus = severityToStatus[severity];
