+++ love/src/components/Watcher/AlarmsTable/AlarmsTable.jsx
@@ -2,7 +2,8 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import Button from '../../GeneralPurpose/Button/Button';
 import RowExpansionIcon from '../../icons/RowExpansionIcon/RowExpansionIcon';
-import Alarm, { severityToStatus } from '../Alarm/Alarm';
+import Alarm from '../Alarm/Alarm';
+import { severityToStatus } from '../../../Config';
 import ColumnHeader from './ColumnHeader/ColumnHeader';
 import DetailsPanel from './DetailsPanel/DetailsPanel';
 import { formatTimestamp, relativeTime } from '../../../Utils';
